US02: Creating the User Interface - Frontend & Backend
"Creating a simple interface. I decided to create a simple header. The header contains a slider for both the Maze Width and Maze Depth, along with numerics on the right of the slider representing the current value of the slider. This is to also in advance keep the
boundaries of US03 in mind where the minimum and maximum width and depth of the maze have been defined as 10x10 and 250x250."

"For today, I made the functionality of the sliders work, where it updates the numbers next to it to show it's value. I also created a Generate Maze Button. This now starts the StartMaze Coroutine, which makes the maze based of the values put into the sliders. The
spawning of the grid is now not so good. It words perfectly the first time you click on it, but then due to how we Instantieate, it becomes a mess after that. For this I will try to implement Object Pooling tomorrow."