Bug Fixed!
Found it! the call to the ClearWalls method only occured when there was a nextCell. So when the algorythm approaches the final cell, and doesn't have a nextCell, the walls don't clear and leaves them at a square. By placing this method outside of the if statements for the activeCells, this gets remedied. now the maze truly generates perfectly and makes it meet the MVP. In total, I needed 12.8 hours to realise this.